Created the 'formatter.py' file
- Created the 'formatter.py' file, to include classes useful to format dataframes, in order to improve the code organization
